requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
aiogram_calendar==0.1.1